feat: useConnectorNameAndIcon hook
